requirements.txt
autopep8==1.7.0
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
pycodestyle==2.9.1
python-decouple==3.6
python-dotenv==0.21.0
six==1.16.0
toml==0.10.2
Werkzeug==2.2.2
